requirements.txt
asgiref==3.6.0
autopep8==2.0.2
boto3==1.26.89
botocore==1.29.89
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
Django==4.1.7
django-colorfield==0.8.0
django-social-share==2.3.0
idna==3.4
jmespath==1.0.1
Pillow==9.4.0
pycodestyle==2.10.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
